split metadata into separate files
